pyproject.toml
[project]
name = "comics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "openai-agents>=0.0.15",
    "openai>=1.77.0",
    "pillow>=11.2.1",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.0",
]
